dev_requirements.txt
asgiref==3.3.1
attrs==20.3.0
beautifulsoup4==4.9.3
coverage==5.5
crispy-tailwind==0.3.0
Django==3.1.6
django-cleanup==5.1.0
django-crispy-forms==1.11.1
django-extensions==3.1.1
django-filter==2.4.0
django-phonenumber-field==5.0.0
django-webtest==1.9.7
djangorestframework==3.12.2
drf-extra-fields==3.1.1
et-xmlfile==1.0.1
flake8==3.8.4
graphviz==0.16
iniconfig==1.1.1
jdcal==1.4.1
mccabe==0.6.1
openpyxl==3.0.6
packaging==20.9
phonenumbers==8.12.18
Pillow==8.1.2
pkg-resources==0.0.0
pluggy==0.13.1
py==1.10.0
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==2.4.7
pytest==6.2.2
pytest-cov==2.11.1
pytz==2021.1
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
toml==0.10.2
waitress==1.4.4
WebOb==1.8.6
WebTest==2.0.35
